+++ tests/src/ArrayValidatorFillerTest.php
@@ -28,7 +28,7 @@ class ArrayValidatorFillerTest extends TestCase
             false
         );
 
-        $value = (new Filler())
+        $value = DefaultFillerRepositoryFactory::create()
             ->fill($validator);
 
         $c = array_diff($value, $enum);
@@ -53,7 +53,7 @@ class ArrayValidatorFillerTest extends TestCase
             true
         );
 
-        $value = (new Filler())
+        $value = DefaultFillerRepositoryFactory::create()
             ->fill($validator);
 
         /*
@@ -79,7 +79,29 @@ class ArrayValidatorFillerTest extends TestCase
             null
         );
 
-        (new Filler())
+        DefaultFillerRepositoryFactory::create()
             ->fill($validator);
+    }
+
+    public function testArrayValidatorWithArrayItemsShouldWork()
+    {
+        $validator = new ArrayValidator(
+            1,
+            1,
+            new ArrayValidator(
+                1,
+                1,
+                new EnumValidator(['1']),
+                false
+            ),
+            false
+        );
+
+        $value = DefaultFillerRepositoryFactory::create()
+            ->fill($validator);
+
+        $this->assertInternalType('array', $value);
+        $this->assertInternalType('array', $value[0]);
+        $this->assertEquals('1', $value[0][0]);
     }
 }