+++ tests/src/FillerRepositoryTest.php
@@ -27,11 +27,11 @@ use Phramework\Validate\StringValidator;
  * @author Xenofon Spafaridis <nohponex@gmail.com>
  * @coversDefaultClass Phramework\ValidateFiller\Filler
  */
-class FillerTest extends TestCase
+class FillerRepositoryTest extends TestCase
 {
-    public function testFillUnknown()
+    public function testFillUnknownValidator()
     {
-        $value = (new Filler())
+        $value = DefaultFillerRepositoryFactory::create()
             ->fill(
                 new DateValidator()
             );
@@ -50,7 +50,7 @@ class FillerTest extends TestCase
             ]
         ];
 
-        $value = (new Filler())
+        $value = DefaultFillerRepositoryFactory::create()
             ->fill(
                 (new ObjectValidator(
                     (object) [